# doc/api.yaml
openapi: '3.0.2'
info:
  title: WASAPhoto API Specifications
  description: |-
    Keep in touch with your friends by sharing photos of special moments, thanks to WASAPhoto! 
    You can upload your photos directly from your PC, and they will be visible to everyone following you.
  version: "1.0"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for authentication in WASAPhoto.
      bearerFormat: '^[a-z0-9]{64}$'
  schemas:
    ID:
      title: ID
      description: |-
        Autoincrement integer field used for identify a particular resource (a comment, a post, ecc.)
      type: integer
      example: 795739
    Username:
      title: Username
      description: Username of a user
      type: string
      example: "chri_genna02"
      minLength: 8
      maxLength: 16
      pattern: '^[a-zA-Z0-9_-]{8,16}$'
    User:
      title: User
      description: |-
        A user in WASAPhoto is represented by a unique ID, its unique username, name and birthdate.
      properties:
        uuid:
          $ref: '#/components/schemas/ID'
        username:
          $ref: '#/components/schemas/Username'
        profilePic:
          $ref: "#/components/schemas/Photo"
        name:
          description: Real-life name/surname of the user
          type: string
          pattern: ''
          minLength: 2
          maxLength: 32
          example: Christian Gennarelli 
        birthdate:
          $ref: '#/components/schemas/Date'
    UserList:
      title: UserList
      description: |-
        Collection of users. 
        A maximum number of 16 users is returned, according to some backend principle (example: datetime)
      type: array
      items:
        $ref: '#/components/schemas/User'  
      minItems: 0
      maxItems: 16 
    Photo:
      title: Photo
      format: binary
      type: string
      description: |-
        Binary representation of a photo, wrapped in a base64-encoded string. 
        Only multiple profile pics are sent via base64 for efficiency reasons.
        URL pointing to post images will be first sent, and then the client will have the ability
        to obtain the photo via a separate endpoint, i.e. GetPhotoFromURL.
      minLength: 0
      maxLength: 2000000 # 2MB photo
      example: iVBORw0KGgbW8gY2hyaXN0aWFuCghyaXN0aWFuCgw0KGgbW8gYyaXN0aWFuCgw0KGgbW8gW8gYyaXN0aW8gYyaXN0aWF
...
    Date:
      title: date
      description: Date.
      example: "2023-11-29"
      type: string
      format: date
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01])$'
      minLength: 10
      maxLength: 10  
    Datetime:
      title: datetime
      description: |-
        Date and time.
      example: "2023-11-29 23:00:00"
      type: string
      format: date-time
      pattern: '^([0-9]{4})-(0[1-9]|1[0-2])-(0[1-9]|[12][0-9]|3[01]) (0[0-9]|1[0-9]|2[0-3]):([0-5][0-9]):([0-5][0-9])$'
      minLength: 19
      maxLength: 19  
    Error:
      title: Error
      description: |-
        Schema defining what is an error when returned in response of a HTTP requests.
      properties:
        code: # Code returned by the server
          type: string
          description: Status code returned directly by the server(s).
          example: "404"
          minLength: 3
          maxLength: 3
        description:  
          type: string
          description: |- 
            Description of the error in human-readable language.
          example: 'Username not found. Please check that the username is correct, and try again.'
          minLength: 1
          maxLength: 128
          pattern: ''
paths:
...
  /users/{username}/banned/:
    parameters:
        - in: path
          name: username
          schema:
            $ref: '#/components/schemas/Username'
          required: true

    get:
      operationId: getBanUserList
      tags: ['BAN']
      summary: Return user banlist.
      description: |-
        Return the list of banned users for the specified user.
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: datetime
          schema:
            $ref: '#/components/schemas/Datetime'
          required: false
      responses:
        '201': # OK
          description: The list of banned users is successfully returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserList'
        '204':
          description: The user has banned no other user.
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client is NOT authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user cannot see banned users of some other user.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put: 
      operationId: banUser
      tags: ['BAN']
      summary: Ban a new user
      description: |-
        Add a new user - here, 'ban_username' - in the list of the banned users for 'username'.
      parameters:
        - in: query
          name: banned_username
          description: Username of the user that must be added to the list of the banned users for this user.
          schema:
            $ref: '#/components/schemas/Username'
          required: true
      security:
        - BearerAuth: []
      responses:
        '204': # OK Created
          description: User 'banned_username' has been successfully added to the ban list.
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client has not been authenticated (yet).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user cannot ban another user on behalf of some other user.
            That is, the authenticated user's username and banned_username do NOT coincide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404': # Content not found
          description: Either one of the two usernames has not been found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{username}/banned/{banned_username}:
    parameters:
        - in: path
          name: username
          schema:
            $ref: '#/components/schemas/Username'
          required: true
        - in: path
          name: banned_username
          schema:
            $ref: '#/components/schemas/Username'
          required: true
    
    delete: 
      operationId: unbanUser
      tags: ['BAN']
      summary: Unban an user
      description: |-
        Delete a user - here, 'banned_username' - from the banned list of the user 'username'.
      security:
        - BearerAuth: []
      responses:
        '204': # OK
          description: User successfully removed banned_username from its ban list.
        '400': # Bad request
          description: Bad request provided.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401': # Unauthenticated
          description: The client has not been authenticated (yet).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403': # Unauthorized
          description: |-
            The authenticated user cannot unban another user on behalf of some other user.
            That is, the authenticated user's username and the one provided in the path do NOT coincide.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500': # Internal server error.
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...